HADOOP-16794  tests for big files.
