+++ hadoop-tools/hadoop-aws/src/test/java/org/apache/hadoop/fs/s3a/AbstractS3ATestBase.java
@@ -18,6 +18,8 @@
 
 package org.apache.hadoop.fs.s3a;
 
+import org.apache.commons.codec.digest.DigestUtils;
+import org.apache.commons.net.util.Base64;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.FileSystem;
 import org.apache.hadoop.fs.Path;
@@ -32,8 +34,11 @@ import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 
+import com.amazonaws.services.s3.model.ObjectMetadata;
+
 import static org.apache.hadoop.fs.contract.ContractTestUtils.dataset;
 import static org.apache.hadoop.fs.contract.ContractTestUtils.writeDataset;
+import static org.apache.hadoop.fs.s3a.Constants.SERVER_SIDE_ENCRYPTION_KEY;
 import static org.apache.hadoop.fs.s3a.S3ATestUtils.getTestDynamoTablePrefix;
 
 /**
@@ -41,6 +46,10 @@ import static org.apache.hadoop.fs.s3a.S3ATestUtils.getTestDynamoTablePrefix;
  */
 public abstract class AbstractS3ATestBase extends AbstractFSContractTestBase
     implements S3ATestConstants {
+
+  protected static final String AWS_KMS_SSE_ALGORITHM = "aws:kms";
+
+  protected static final String SSE_C_ALGORITHM = "AES256";
 
   protected static final Logger LOG =
       LoggerFactory.getLogger(AbstractS3ATestBase.class);
@@ -159,4 +168,63 @@ public abstract class AbstractS3ATestBase extends AbstractFSContractTestBase
       throw e;
     }
   }
+
+  /**
+   * Assert that a path is encrypted with right encryption settings.
+   * @param path file path.
+   * @param algorithm encryption algorithm.
+   * @param kmsKeyArn
+   * @throws IOException
+   */
+  protected void assertEncrypted(final Path path,
+                                 final S3AEncryptionMethods algorithm,
+                                 final String kmsKeyArn)
+          throws IOException {
+    ObjectMetadata md = getFileSystem().getObjectMetadata(path);
+    String details = String.format(
+            "file %s with encryption algorthm %s and key %s",
+            path,
+            md.getSSEAlgorithm(),
+            md.getSSEAwsKmsKeyId());
+    switch(algorithm) {
+      case SSE_C:
+        assertNull("Metadata algorithm should have been null in "
+                        + details,
+                md.getSSEAlgorithm());
+        assertEquals("Wrong SSE-C algorithm in "
+                        + details,
+                SSE_C_ALGORITHM, md.getSSECustomerAlgorithm());
+        String md5Key = convertKeyToMd5();
+        assertEquals("getSSECustomerKeyMd5() wrong in " + details,
+                md5Key, md.getSSECustomerKeyMd5());
+        break;
+      case SSE_KMS:
+        assertEquals("Wrong algorithm in " + details,
+                AWS_KMS_SSE_ALGORITHM, md.getSSEAlgorithm());
+        assertEquals("Wrong KMS key in " + details,
+                kmsKeyArn,
+                md.getSSEAwsKmsKeyId());
+        break;
+      default:
+        assertEquals("AES256", md.getSSEAlgorithm());
+    }
+  }
+
+  /**
+   * Decodes the SERVER_SIDE_ENCRYPTION_KEY from base64 into an AES key, then
+   * gets the md5 of it, then encodes it in base64 so it will match the version
+   * that AWS returns to us.
+   *
+   * @return md5'd base64 encoded representation of the server side encryption
+   * key
+   */
+  private String convertKeyToMd5() {
+    String base64Key = getFileSystem().getConf().getTrimmed(
+            SERVER_SIDE_ENCRYPTION_KEY
+    );
+    byte[] key = Base64.decodeBase64(base64Key);
+    byte[] md5 =  DigestUtils.md5(key);
+    return Base64.encodeBase64String(md5).trim();
+  }
+
 }
