+++ hadoop-tools/hadoop-aws/src/test/java/org/apache/hadoop/fs/s3a/AbstractTestS3AEncryption.java
@@ -20,11 +20,8 @@ package org.apache.hadoop.fs.s3a;
 
 import java.io.IOException;
 
-import com.amazonaws.services.s3.model.ObjectMetadata;
 import org.junit.Test;
 
-import org.apache.commons.codec.digest.DigestUtils;
-import org.apache.commons.net.util.Base64;
 import org.apache.hadoop.conf.Configuration;
 import org.apache.hadoop.fs.Path;
 import org.apache.hadoop.fs.contract.ContractTestUtils;
@@ -43,10 +40,6 @@ import static org.apache.hadoop.fs.s3a.S3AUtils.getEncryptionAlgorithm;
  * file length may be rounded up to match word boundaries.
  */
 public abstract class AbstractTestS3AEncryption extends AbstractS3ATestBase {
-
-  protected static final String AWS_KMS_SSE_ALGORITHM = "aws:kms";
-
-  protected static final String SSE_C_ALGORITHM = "AES256";
 
   @Override
   protected Configuration createConfiguration() {
@@ -169,57 +162,6 @@ public abstract class AbstractTestS3AEncryption extends AbstractS3ATestBase {
         getTrimmed(SERVER_SIDE_ENCRYPTION_KEY);
     S3AEncryptionMethods algorithm = getSSEAlgorithm();
     assertEncrypted(path, algorithm, kmsKeyArn);
-  }
-
-  public void assertEncrypted(final Path path,
-      final S3AEncryptionMethods algorithm,
-      final String kmsKeyArn)
-      throws IOException {
-    ObjectMetadata md = getFileSystem().getObjectMetadata(path);
-    String details = String.format(
-        "file %s with encryption algorthm %s and key %s",
-        path,
-        md.getSSEAlgorithm(),
-        md.getSSEAwsKmsKeyId());
-    switch(algorithm) {
-    case SSE_C:
-      assertNull("Metadata algorithm should have been null in "
-          + details,
-          md.getSSEAlgorithm());
-      assertEquals("Wrong SSE-C algorithm in "
-          + details,
-          SSE_C_ALGORITHM, md.getSSECustomerAlgorithm());
-      String md5Key = convertKeyToMd5();
-      assertEquals("getSSECustomerKeyMd5() wrong in " + details,
-          md5Key, md.getSSECustomerKeyMd5());
-      break;
-    case SSE_KMS:
-      assertEquals("Wrong algorithm in " + details,
-          AWS_KMS_SSE_ALGORITHM, md.getSSEAlgorithm());
-      assertEquals("Wrong KMS key in " + details,
-          kmsKeyArn,
-          md.getSSEAwsKmsKeyId());
-      break;
-    default:
-      assertEquals("AES256", md.getSSEAlgorithm());
-    }
-  }
-
-  /**
-   * Decodes the SERVER_SIDE_ENCRYPTION_KEY from base64 into an AES key, then
-   * gets the md5 of it, then encodes it in base64 so it will match the version
-   * that AWS returns to us.
-   *
-   * @return md5'd base64 encoded representation of the server side encryption
-   * key
-   */
-  private String convertKeyToMd5() {
-    String base64Key = getFileSystem().getConf().getTrimmed(
-        SERVER_SIDE_ENCRYPTION_KEY
-    );
-    byte[] key = Base64.decodeBase64(base64Key);
-    byte[] md5 =  DigestUtils.md5(key);
-    return Base64.encodeBase64String(md5).trim();
   }
 
   protected abstract S3AEncryptionMethods getSSEAlgorithm();
